help

+++ client/client.py
@@ -2,11 +2,11 @@ from person import Person
 
 
 class Client(Person):
-    def __init__(self, name: str, money: int):
+    def __init__(self, name: str, money: float):
         super().__init__(name)
-        self.money: int = money
+        self.money: float = money
 
-    def get_money(self) -> int:
+    def get_money(self) -> float:
         return self.money
 
     def set_money(self, money: int):
